Full refactor
