+++ result.json
@@ -2,7 +2,7 @@
   "schema_version": "0.4.0",
   "metadata": {
     "name": "near-abi-py",
-    "version": "0.1.0",
+    "version": "0.4.0",
     "authors": [
       "Unknown"
     ],
