+++ src/near_abi_py/__init__.py
@@ -1,5 +1,5 @@
 """
-NEAR Python ABI Builder (Ultra-Simplified)
+NEAR Python ABI Builder
 
 A minimalist tool for generating Application Binary Interface (ABI) definitions
 for NEAR smart contracts written in Python.
@@ -7,7 +7,7 @@ for NEAR smart contracts written in Python.
 
 from .generator import generate_abi, generate_abi_from_files
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
     "generate_abi",
