--- src/near_abi_py/analyzer/__init__.py
@@ -1,13 +1,0 @@
-"""
-Function analysis utilities for NEAR ABI generation.
-"""
-
-from .function import FunctionAnalyzer
-from .inspector import FunctionInspector
-from .loader import ModuleLoader
-
-__all__ = [
-    "FunctionAnalyzer",
-    "FunctionInspector",
-    "ModuleLoader",
-]
